crud complete

+++ TaskManager/Controllers/TasksController.cs
@@ -63,13 +63,48 @@ namespace TaskManager.Controllers
         [HttpDelete("delete")]
         public async Task<JsonResult> DeleteTaskAsync([FromBody] string id)
         {
-            Guid.TryParse(id, out Guid key);
+            if (!Guid.TryParse(id, out Guid key))
+            {
+                return new JsonResult(new { success = false, error = $"Unable to find parse with id: {id}" });
+            }
 
             try
             {
                 Models.Task taskToDelete = await context.Tasks.FindAsync(key);
 
+                if (taskToDelete == null)
+                {
+                    return new JsonResult(new { success = false, error = $"Unable to find Task with id: {id}" });
+                }
+
                 context.Tasks.Remove(taskToDelete);
+                await context.SaveChangesAsync();
+
+                return new JsonResult(new { success = true });
+            }
+            catch (Exception ex)
+            {
+                return new JsonResult(new { success = false, error = ex.Message });
+            }
+        }
+
+        [HttpPut("edit")]
+        public async Task<JsonResult> EditTaskAsync([FromBody]  Models.Task task)
+        {
+            try
+            {
+                Models.Task existingRecord = await context.Tasks.FindAsync(task.Id);
+
+                if (existingRecord == null)
+                {
+                    return new JsonResult(new { success = false, error = $"Unable to find Task with id: {task.Id}" });
+                }
+
+                existingRecord.Title = task.Title;
+                existingRecord.Description = task.Description;
+                existingRecord.DueDate = task.DueDate;
+
+                context.Tasks.Update(existingRecord);
                 await context.SaveChangesAsync();
 
                 return new JsonResult(new { success = true });
